PM-10069 - Refactor unauthExtensionRefreshSwap to be available in non-extension clients and rename.
